pyproject.toml
[project]
name = "kaggle-wheat-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "albumentations>=2.0.8",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
]
